+++ rustendo_lib/src/mappers/mapper_000.rs
@@ -17,24 +17,24 @@ impl Mapper000 {
 }
 
 impl Mapper for Mapper000 {
-    fn cpu_read(&self, address: u16) -> (Option<u16>, Option<u8>) {
+    fn cpu_read(&self, address: u16) -> (Option<usize>, Option<u8>) {
         match address {
             // Unused, but in the cartridge's address range
             0x4020..=0x5FFF => (None, None),
             0x6000..=0x7FFF => (None, Some(self.prg_ram[(address & 0x1FFF) as usize])),
-            0x8000..=0xBFFF => (Some(address & 0x7FFF), None),
+            0x8000..=0xBFFF => (Some((address & 0x7FFF) as usize), None),
             0xC000..=0xFFFF => match self.prg_rom_size {
                 // If the size is 16 KiB, mirror
                 0x4000 => self.cpu_read(address & 0xBFFF),
                 // If the size is 32 KiB, continue previous range
-                0x8000 => (Some(address & 0x7FFF), None),
+                0x8000 => (Some((address & 0x7FFF) as usize), None),
                 _ => (None, None),
             },
             _ => (None, None),
         }
     }
 
-    fn cpu_write(&mut self, address: u16, data: u8) -> Option<u16> {
+    fn cpu_write(&mut self, address: u16, data: u8) -> Option<usize> {
         match address {
             0x6000..=0x7FFF => {
                 self.prg_ram[(address & 0x1FFF) as usize] = data;
@@ -44,17 +44,17 @@ impl Mapper for Mapper000 {
         }
     }
 
-    fn ppu_read(&self, address: u16) -> (Option<u16>, Option<u8>) {
+    fn ppu_read(&self, address: u16) -> (Option<usize>, Option<u8>) {
         match address {
             0x0000..=0x1FFF => match self.chr_ram.len() {
-                0 => (Some(address), None),
+                0 => (Some(address as usize), None),
                 _ => (None, Some(self.chr_ram[address as usize])),
             },
             _ => (None, None),
         }
     }
 
-    fn ppu_write(&mut self, address: u16, data: u8) -> Option<u16> {
+    fn ppu_write(&mut self, address: u16, data: u8) -> Option<usize> {
         match self.chr_ram.len() {
             0 => None,
             _ => {
