+++ rustendo_lib/src/mappers/mod.rs
@@ -1,8 +1,8 @@
 pub trait Mapper {
-    fn cpu_read(&self, address: u16) -> (Option<u16>, Option<u8>);
-    fn cpu_write(&mut self, address: u16, data: u8) -> Option<u16>;
-    fn ppu_read(&self, address: u16) -> (Option<u16>, Option<u8>);
-    fn ppu_write(&mut self, address: u16, data: u8) -> Option<u16>;
+    fn cpu_read(&self, address: u16) -> (Option<usize>, Option<u8>);
+    fn cpu_write(&mut self, address: u16, data: u8) -> Option<usize>;
+    fn ppu_read(&self, address: u16) -> (Option<usize>, Option<u8>);
+    fn ppu_write(&mut self, address: u16, data: u8) -> Option<usize>;
 }
 
 pub mod mapper_000;
